pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vparse"
dynamic = ["version"]
license = { text = "AGPL-3.0" }
description = "A practical tool for converting PDF to Markdown"
readme = "README.md"
requires-python = ">=3.10,<3.14"
keywords = ["magic-pdf", "vparse", "VParse", "convert", "pdf", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "boto3>=1.28.43",
    "click>=8.1.7",
    "loguru>=0.7.2",
    "numpy>=1.21.6",
    "pdfminer.six>=20251230",
    "tqdm>=4.67.1",
    "requests",
    "httpx",
    "pillow>=11.0.0",
    "pypdfium2>=4.30.0",
    "pypdf>=5.6.0",
    "reportlab",
    "pdftext>=0.6.3",
    "json-repair>=0.46.2",
    "opencv-python>=4.11.0.86",
    "fast-langdetect>=0.2.3,<0.3.0",
    "scikit-image>=0.25.0,<1.0.0",
    "openai>=1.70.0,<3",
    "beautifulsoup4>=4.13.5,<5",
    "magika>=0.6.2,<1.1.0",
]

[project.optional-dependencies]
test = [
    "vparse[core]",
    "pytest",
    "pytest-cov",
    "coverage",
    "fuzzywuzzy"
]
dev = [
    "ruff>=0.8.0",
    "vulture>=2.0",
]
tesseract = [
    "pytesseract>=0.3.10",
]
lite = [
    "vparse[tesseract]",
]
vlm = [
    "torch>=2.6.0,<3",
    "transformers>=4.51.1,!=4.57.2,<5.0.0",
    "accelerate>=1.5.1",
    "mineru-vl-utils>=0.1.19.1,<1",
    "qwen-vl-utils>=0.0.14,<1",
    "modelscope>=1.26.0",
    "huggingface-hub>=0.32.4",
]
vllm = [
    "vllm>=0.10.1.1,<0.12",
]
lmdeploy = [
    "lmdeploy>=0.10.2,<0.12",
]
mlx = [
    "mlx-vlm>=0.3.3,<0.4",
]
pipeline = [
    "matplotlib>=3.10,<4",
    "ultralytics>=8.3.48,<9",
    "doclayout_yolo==0.0.4",
    "dill>=0.3.8,<1",
    "PyYAML>=6.0.1,<7",
    "ftfy>=6.3.1,<7",
    "shapely>=2.0.7,<3",
    "pyclipper>=1.3.0,<2",
    "omegaconf>=2.3.0,<3",
    "torch>=2.6.0,<3",
    "torchvision",
    "transformers>=4.49.0,!=4.51.0,<5.0.0",
    "onnxruntime>1.17.0",
    "modelscope>=1.26.0",
    "huggingface-hub>=0.32.4",
]
api = [
    "fastapi",
    "python-multipart",
    "uvicorn",
]
gradio = [
    "gradio==5.49.1",
    "gradio-pdf==0.0.22",
]
core = [
    "vparse[lite]",
    "vparse[vlm]",
    "vparse[pipeline]",
    "vparse[api]",
    "vparse[gradio]",
]
all = [
    "vparse[core]",
    "vparse[mlx] ; sys_platform == 'darwin'",
    "vparse[vllm] ; sys_platform == 'linux'",
    "vparse[lmdeploy] ; sys_platform == 'windows'",
]

[project.urls]
homepage = "https://vparse.net/"
documentation = "https://opendatalab.github.io/VParse/"
repository = "https://github.com/opendatalab/VParse"
issues = "https://github.com/opendatalab/VParse/issues"

[project.scripts]
vparse = "vparse.cli.__main__:main"
vparse-vllm-server = "vparse.cli.vlm_server:vllm_server"
vparse-lmdeploy-server = "vparse.cli.vlm_server:lmdeploy_server"
vparse-openai-server = "vparse.cli.vlm_server:openai_server"
vparse-models-download = "vparse.cli.models_download:download_models"
vparse-api = "vparse.cli.fast_api:main"
vparse-gradio = "vparse.cli.gradio_app:main"
# Backward-compatible CLI aliases
mineru = "vparse.cli.__main__:main"
mineru-vllm-server = "vparse.cli.vlm_server:vllm_server"
mineru-lmdeploy-server = "vparse.cli.vlm_server:lmdeploy_server"
mineru-openai-server = "vparse.cli.vlm_server:openai_server"
mineru-models-download = "vparse.cli.models_download:download_models"
mineru-api = "vparse.cli.fast_api:main"
mineru-gradio = "vparse.cli.gradio_app:main"

[tool.setuptools.dynamic]
version = { attr = "vparse.version.__version__" }

[tool.setuptools.packages.find]
include = ["vparse*", "mineru*"]
namespaces = false

[tool.setuptools.package-data]
"vparse" = ["resources/**", "py.typed"]
"vparse.model.utils.pytorchocr.utils" = ["resources/**"]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.pytest.ini_options]
addopts = "-s --cov=vparse --cov-report html"

[tool.coverage.run]
command_line = "-m pytest tests/unittest/test_e2e.py"
source = ["vparse/"]
omit = [
    "*/gradio_app.py",
    "*/models_download.py",
    "*/fast_api.py",
    "*/cli/client.py",
    "*/vlm_vllm_server.py",
    "*/cli_parser.py",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.report]
exclude_also = [
    'def __repr__',
    'if self.debug:',
    'if settings.DEBUG',
    'raise AssertionError',
    'raise NotImplementedError',
    'if 0:',
    'if __name__ == .__main__.:',
    'if TYPE_CHECKING:',
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
]
